Created the wrapper to query for existing components in entities

+++ Editor/Source/Components/ImguiDialog.cpp
@@ -29,7 +29,7 @@ void ImguiDialog::submit(graphics::Renderer2D* renderer) const
 		static int selection_mask = (1 << 2); // Dumb representation of what may be user-side selection state. You may carry selection state inside or outside your objects in whatever format you see fit.
 		int node_clicked = -1;                // Temporary storage of what node we have clicked to process selection at the end of the loop. May be a pointer to your own node type, etc.
 		ImGui::PushStyleVar(ImGuiStyleVar_IndentSpacing, ImGui::GetFontSize() * 3); // Increase spacing to differentiate leaves from expanded contents.
-		for (int i = 0; i < entities.size(); i++)
+		for (int i = 0; i < (int)entities.size(); i++)
 		{
 			// Disable the default open on single-click behavior and pass in Selected flag according to our selection state.
 			ImGuiTreeNodeFlags node_flags = ImGuiTreeNodeFlags_OpenOnArrow | ImGuiTreeNodeFlags_OpenOnDoubleClick | ((selection_mask & (1 << i)) ? ImGuiTreeNodeFlags_Selected : 0);
